Implementing the UI
I implemented a basic UI that contains a title and a button.

Fixing comments

* I moved main.dart to lib folder.
* I removed an unuseless class (dice_dimens) because in dont use constants in code.
* I removed the StreamBuilder without use.
* I changed all the message of the last commit, spanish to english
  commit.
* I created a separate file to app widget.
* I added the necessary trailing commas on the last parameters to help
  Dart to format and organize the code.
